Create DevCard.yml
There are two ways my action is triggered, the first being if there is a push on the master branch, and the second one is a schedule, which acts as a cronjob and will return every {x} time. In my case, I'll run it every night at 00:00.

Then created a new job that will run the DevCard GitHub action. The variable for this version is the ID of my DevCard which I'm fetching.